made all pertinent  changes from the team meeting feedback
