+++ smart-contracts/contracts/incentive-gauge-factory/src/lib.rs
@@ -5,13 +5,9 @@ pub mod state;
 pub mod msg;
 pub mod contract;
 
-pub mod migrate;
 pub mod queries;
 pub mod executes;
 pub mod replies;
-
-pub mod gauge;
-pub mod fees;
 
 pub mod types;
