+++ smart-contracts/contracts/merkle-incentives/Cargo.toml
@@ -44,6 +44,7 @@ schemars = "0.8.15"
 serde = {version = "1.0.189", default-features = false, features = ["derive"]}
 serde-json-wasm = "1.0.1"
 thiserror = {version = "1.0.49"}
+semver = "1.0.20"
 
 [dev-dependencies]
 serde_json = "1.0"
