+++ smart-contracts/contracts/merkle-incentives/src/msg.rs
@@ -32,4 +32,8 @@ pub enum QueryMsg {
 }
 
 #[cw_serde]
-pub struct MigrateMsg {}
+pub struct MigrateMsg {
+    pub version: String,
+}
+
+
